% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_dplyr.R
\name{calculate_summary_dplyr}
\alias{calculate_summary_dplyr}
\title{Calculate summary function}
\usage{
calculate_summary_dplyr(input_dataframe)
}
\arguments{
\item{input_dataframe:}{grouped dataframe}
}
\description{
This function calculates mean, SD, SEM, and limits (mean +/- SEM) on a variable (value) within a dataframe (input_dataframe).
Returns global variable `summarized_data`
}

