% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_holm_bonferroni.R
\name{compute_holm_bonferroni}
\alias{compute_holm_bonferroni}
\title{Bonferroni-Holm correction
for published example, see: https://www.pnas.org/content/pnas/111/40/E4264.full.pdf
online tutorial: https://www.statisticshowto.datasciencecentral.com/holm-bonferroni-method/
theoretically, should be able to use `p.adjust`, but results are non-sensical
based on: https://rcompanion.org/rcompanion/f_01.html
1. rank order the p-values
2. divide the largest tolerated p-value (alpha value) by the rank - this is the crux of the Holm correction
3. return whether or not the observed p-value is smaller (and, thus, significant) than the Holm corrected p-value
TODO: Figure out how to dynamically set name of computed column so that contains name of `pvalue_column` (think: the way that `summarise` works)}
\usage{
compute_holm_bonferroni(input_data, pvalue_column, alpha_threshold)
}
\description{
Bonferroni-Holm correction
for published example, see: https://www.pnas.org/content/pnas/111/40/E4264.full.pdf
online tutorial: https://www.statisticshowto.datasciencecentral.com/holm-bonferroni-method/
theoretically, should be able to use `p.adjust`, but results are non-sensical
based on: https://rcompanion.org/rcompanion/f_01.html
1. rank order the p-values
2. divide the largest tolerated p-value (alpha value) by the rank - this is the crux of the Holm correction
3. return whether or not the observed p-value is smaller (and, thus, significant) than the Holm corrected p-value
TODO: Figure out how to dynamically set name of computed column so that contains name of `pvalue_column` (think: the way that `summarise` works)
}
