% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_trailing_slash.R
\name{ensure_trailing_slash}
\alias{ensure_trailing_slash}
\title{Function to ensure path ends in a trailing slash}
\usage{
ensure_trailing_slash(path)
}
\arguments{
\item{path}{Input string to check for a trailing slash}
}
\value{
path Input string with a trailing slash (if had at slash at time of input, will not duplicate)
}
\description{
Function to ensure path ends in a trailing slash
}

