% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_r_to_z.R
\name{fisher_r_to_z}
\alias{fisher_r_to_z}
\title{Fisher's r to z transform
//TODO investigate using fisherz function instead}
\usage{
fisher_r_to_z(data)
}
\arguments{
\item{data:}{column from a dataframe to do fishers transform on}
}
\description{
Fisher's r to z transform
//TODO investigate using fisherz function instead
}
