% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fml.R
\name{fml}
\alias{fml}
\title{FML function}
\usage{
fml()
}
\description{
based on:
http://stackoverflow.com/questions/21718711/writing-a-shortcut-function-in-r-for-rmlist-ls
http://stackoverflow.com/questions/14260340/function-to-clear-the-console-in-r
}
\details{
Clears the environment (like rm(list=ls()))
Clears the screen (like ctrl+L)
}

