% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_three_digit_subject_id.R
\name{format_three_digit_subject_id}
\alias{format_three_digit_subject_id}
\title{Format subject IDs to s### format}
\usage{
format_three_digit_subject_id(sub)
}
\arguments{
\item{sub}{Subject number}
}
\value{
formatted subject number
}
\description{
Format subject IDs to s### format
}
