% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_s_to_subject_id.R
\name{prepend_s_to_subject_id}
\alias{prepend_s_to_subject_id}
\title{Format subject IDs by appending s to subjectID so that format is s#}
\usage{
prepend_s_to_subject_id(sub)
}
\arguments{
\item{sub}{Subject number}
}
\value{
formatted subject number
}
\description{
Format subject IDs by appending s to subjectID so that format is s#
}
