% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_NA_with_zeros.R
\name{replace_NA_with_zeros}
\alias{replace_NA_with_zeros}
\title{Function to replace NAs with 0s}
\usage{
replace_NA_with_zeros(x)
}
\arguments{
\item{x}{Input data that may have NAs}
}
\description{
based on: http://stackoverflow.com/questions/26081672/change-variable-values-by-groups-using-dplyr
}

