% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cohens_d.R
\name{compute_cohens_d}
\alias{compute_cohens_d}
\title{Cohen's d function}
\usage{
compute_cohens_d(input_df)
}
\description{
Compute an effect size (Cohen's d) for paired samples t-tests.
To use, your dataframe must have just 2 columns called `var1` and `var2` that correspond to the two conditions of interest.
These columns must be paired data of the same length such that each row represents observations from a single subject.
To call function:
compute_cohens_d(data_frame_with_var1_var2_columns)
}
\details{
It will return three versions of Cohen's d
Standard: mean / sd

Morris DeShon: Computes effect size while correcting for the correlation between condition means
Generally takes the form of `mean_difference / (mean_sd * (sqrt(2*(1-correlation))))`
This can be found in their paper under Equation 8.
Morris, S. B., & DeShon, R. P. (2002). Combining effect size estimates in meta-analysis with repeated measures and independent-groups designs. Psychological Methods, 7(1), 105–125. http://doi.org/10.1037//1082-989X.7.1.105

Dunlap: Computes effect size while correcting for the correlation between condition means. Uses the t-value rather than computing the difference between the means as is done with Morris DeShon.
Generally takes the form of `t_statistic * (((2 *(1 - correlation)) / (number_of_subjects) ^ 0.5)`
This can be found in their paper under Equation 3.
Dunlap, W. P., Cortina, J. M., Vaslow, J. B., & Burke, M. J. (1996). Meta-analysis of experiments with matched groups or repeated measures designs. Psychological Methods, 1(2), 170–177. http://doi.org/10.1037/1082-989X.1.2.170

If you'd like to verify this function works with a set of test data, try one of the following.
Test data courtesy of Tanya Jonker (tanyarjonker@gmail.com) and results based on calculations from her Excel workbook.
Example 1 test data
e1_condA <- c(0.50,0.88,0.56,0.69,1.00,0.56,0.69,0.56,0.81,0.69,0.69,0.63,0.56,0.69,0.81,0.81,0.63,0.75,0.88,0.69,0.75,0.69,0.69,0.69,0.56,0.81,0.56,0.69,0.75,0.63)
e1_condB <- c(0.75,1.00,0.75,0.75,0.94,0.69,0.44,0.69,0.81,0.56,0.75,0.69,0.56,0.88,0.94,0.50,0.69,0.75,1.00,0.69,0.94,0.50,0.63,0.75,0.31,0.81,0.94,0.81,0.81,0.75)
e1_df <- data.frame(e1_condA, e1_condB)
TJ would estimate the effect size as
standard: -0.259328434
Morris DeShon: -0.271365019
Dunlap: -0.268558307

example 2 test data
e2_condA <- c(0.38,0.63,0.69,0.50,0.88,0.50,0.75,0.81,0.56,0.31,0.69,0.56,0.38,0.69,0.44,0.38,0.50,0.69,0.88,0.75,0.75,0.56,0.94,0.63,0.44,0.69,0.81,0.69,0.69,0.69)
e2_condB <- c(0.56,0.94,0.88,0.56,0.81,0.50,0.50,0.75,0.69,0.38,0.81,0.31,0.69,0.69,0.63,0.56,0.63,0.75,0.94,0.75,0.75,0.75,0.88,0.44,0.50,1.00,0.75,0.75,0.88,0.94)
e2_df <- data.frame(e2_condA, e2_condB)
TJ would estimate the effect size as
standard: -0.468829054
Morris DeShon: -0.470087261
Dunlap: -0.413884462
}

